+++ iOSswiftUIa11yTechniques/iOSswiftUIa11yTechniques/ContentView.swift
@@ -163,6 +163,7 @@ struct ContentView: View {
         case "maps": MapView()
         case "meaningful accessible names": MeaningfulAccessibleNamesView()
         case "menus": MenusView()
+        case "multi-selection lists": MultiSelectionListView()
         case "navigation": NavigationLinkView()
         case "page titles": PageTitlesView()
         case "pickers": PickersView()